% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EEG_Eye_State}
\alias{EEG_Eye_State}
\title{EEG Eye State Data Set}
\format{
14,980 x 15 matrix. First fourteen columns are V1 - V14, final is eyeDetection.
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/EEG+Eye+State#}
}
\usage{
EEG_Eye_State
}
\description{
All data is from one continuous EEG measurement with the Emotiv EEG Neuroheadset. The duration of the measurement was 117 seconds. The eye state was detected via a camera during the EEG measurement and added later manually to the file after analysing the video frames. '1' indicates the eye-closed and '0' the eye-open state. All values are in chronological order with the first measured value at the top of the data.
}
\keyword{datasets}
