% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.classifier.R
\name{Other.classifier}
\alias{Other.classifier}
\title{The user's favourite classifier}
\usage{
Other.classifier(x, grouping, xTest, CV = FALSE, ...)
}
\arguments{
\item{x}{An n by p matrix containing the training dataset.}

\item{grouping}{A vector of length n containing the training data classes.}

\item{xTest}{An n.test by p test dataset.}

\item{CV}{If TRUE perform cross-validation (or otherwise) to classify training set. If FALSE, classify test set.}

\item{...}{Optional arguments e.g. tuning parameters}
}
\value{
A vector of classes of the training or test set.
}
\description{
User defined code to convert existing R code for classification to the correct format.
}
\details{
User editable code for your choice of base classifier.
}
\examples{
Other.classifier(NULL, NULL, NULL)
}
