% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPChoose.R
\name{RPChooseA}
\alias{RPChooseA}
\title{Chooses projection}
\usage{
RPChooseA(
  XTrain,
  YTrain,
  XTest,
  d,
  B2 = 10,
  base = "LDA",
  k = c(3, 5),
  projmethod = "Haar",
  estmethod = "training",
  ...
)
}
\arguments{
\item{XTrain}{An n by p matrix containing the training data feature vectors.}

\item{YTrain}{A vector of length n of the classes (either 1 or 2) of the training data.}

\item{XTest}{An n.test by p matrix of the test data.}

\item{d}{The lower dimension of the image space of the projections.}

\item{B2}{The block size.}

\item{base}{The base classifier: one of "knn","LDA","QDA" or "other".}

\item{k}{The options for k if base is "knn".}

\item{projmethod}{Either "Haar", "Gaussian" or "axis".}

\item{estmethod}{Method for estimating the test errors to choose the projection: either training error "training" or leave-one-out "loo".}

\item{...}{Optional further arguments if base = "other".}
}
\value{
A matrix of size p by d giving the selected projection.
}
\description{
Chooses a the best projection from a set of size B2 based on a test error estimate.
}
\details{
Randomly projects the the data B2 times. Chooses the projection yielding the smallest estimate of the test error.
}
\note{
Resubstitution method unsuitable for the k-nearest neighbour classifier.
}
