% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPParallel.R
\name{RPParallel}
\alias{RPParallel}
\title{Chooses a projection from each block in parallel}
\usage{
RPParallel(
  XTrain,
  YTrain,
  XVal = NULL,
  YVal = NULL,
  XTest,
  d,
  B1 = 500,
  B2 = 50,
  base = "LDA",
  projmethod = "Gaussian",
  estmethod = "training",
  k = c(3, 5, 9),
  clustertype = "Default",
  cores = 1,
  machines = NULL,
  seed = 1,
  ...
)
}
\arguments{
\item{XTrain}{An n by p matrix containing the training data feature vectors.}

\item{YTrain}{A vector of length n containing the classes (either 1 or 2) of the training data.}

\item{XVal}{An n.val by p matrix containing the validation data feature vectors.}

\item{YVal}{A vector of length n.val of the classes (either 1 or 2) of the validation data.}

\item{XTest}{An n.test by p matrix containing the test data feature vectors.}

\item{d}{The lower dimension of the image space of the projections.}

\item{B1}{The number of blocks.}

\item{B2}{The size of each block.}

\item{base}{The base classifier one of "knn","LDA","QDA" or "other".}

\item{projmethod}{"Haar", "Gaussian" or "axis".}

\item{estmethod}{Method for estimating the test errors to choose the projection: either training error "training", leave-one-out "loo", or sample split "samplesplit".}

\item{k}{The options for k if base is "knn".}

\item{clustertype}{The type of cluster: "Default" uses just one core, "Fork" uses a single machine, "Socket" uses many machines. Note "Fork" and "Socket" are not supported on windows.}

\item{cores}{Required only if clustertype==Fork: the number of computer cores to use (note: cores > 1 not supported on Windows).}

\item{machines}{Required only if clustertype==Socket: the names of the machines to use e.g. c("Computer1", "Computer2") (not supported on Windows).}

\item{seed}{If not NULL, sets random seed for reproducible results.}

\item{...}{Optional further arguments if base = "other".}
}
\value{
If estmethod == "training" or "loo" , then returns an n+n.test by B1 matrix, each row containing the result of a call to RPChoose. If estmethod == "samplesplit", then returns an n.val+n.test by B1 matrix, each row containing the result of a call to RPChooseSS.
}
\description{
Makes B1 calls to RPChoose or RPChooseSS in parallel and returns the results as a matrix.
}
\details{
Makes B1 calls to RPChoose or RPChooseSS in parallel..
}
\examples{
Train <- RPModel(1, 50, 100, 0.5)
Test <- RPModel(1, 100, 100, 0.5)
Out <- RPParallel(XTrain = Train$x, YTrain = Train$y, XTest = Test$x, d = 2, B1 = 10, 
  B2 = 10, base = "LDA", projmethod = "Haar", estmethod = "training")
colMeans(Out)
}
\seealso{
RPChoose, RPChooseSS.
}
