% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPModel.R
\name{RPModel}
\alias{RPModel}
\title{Generate pairs x,y from joint distribution}
\usage{
RPModel(Model.No, n, p, Pi = 0.5)
}
\arguments{
\item{Model.No}{Model Number}

\item{n}{Sample size}

\item{p}{Data dimension}

\item{Pi}{Class one prior probability}
}
\value{
\describe{
  \item{x}{An n by p data matrix – n observations of the p-dimensional features.}
  \item{y}{A vector of length n containing the classes (either 1 or 2).}
  }
}
\description{
Generates data from the models described in Cannings and Samworth (2016).
}
\note{
Models 1 and 2 require p = 100 or 1000.
}
\examples{
Data <- RPModel(Model.No = 1, 100, 100, Pi = 1/2)
table(Data$y)
colMeans(Data$x[Data$y==1,])
colMeans(Data$x[Data$y==2,])
}
